Move `organizations` to `features` directory
